pyproject.toml
[project]
name = "unacovskaya-online-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.25.0",
    "django>=6.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "vkbottle>=4.6.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
